requirements.txt
PyQt5==5.15.10
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
openslide-python==1.3.1
scikit-image==0.22.0
